+++ RELEASE-NOTES.md
@@ -57,6 +57,7 @@ class Args {
 - #195 Usage help should show Map types if paramLabel not specified
 - #185 Missing option exception text should not use field names but be more descriptive and consistent with usage help. Thanks to [AlexFalappa](https://github.com/AlexFalappa). 
 - #186 Confusing usage message for collection options
+- #179 Remove full JRE dependency: require only Compact Profile. Replace use of `java.awt.Point` with `picocli.CommandLine.Help.TextTable.Cell`. Thanks to [webfolderio](https://github.com/webfolderio).
 - #181 Fixed bug where incorrect help message is displayed for short options with paramLabel when arity > 1
 - #184 Improved CommandLine.setSeparator javadoc to clarify that this affects parsing only and link to the `@Command` `separator` annotation attribute.